// backend/src/domain/models/userOld.ts
export interface UserOldModel {
  id?: number;
  realm: string;
  username: string;
  nomeUsuario: string;
  email: string;
  emailVerified: boolean;
  verificationToken: string;
  idiomaUsuario: string;
  modelo: string;
  temperaturaUsuario: string;
  dataCadastro: string;
  tipoUsuario: string;
  statusUsuario: string;
}